# Remove OPA filter `dependency.base` calculation in optional body parsing