Due to a detected performance concern for CPU as seen below,

<img width="1440" alt="image" src="https://github.com/user-attachments/assets/f2f4991d-47d1-4897-bcda-c524421555ad">

**Memory profile**

_alloc_objects profile_
<img width="1440" alt="image" src="https://github.com/user-attachments/assets/a8d2faca-e171-46f5-8c22-82e38621253a">


### Load test results
<details>
<summary>Using a request body as below with a simple policy to check email domain.</summary>

```
{
      "email": "loadtest@zalando.de",
      "user": {
        "id": "12345",
        "name": {
          "first_name": "John",
          "last_name": "Doe"},
        "contact": {
          "phone_number": "+49-123-456-7890",
          "address": {
            "street": "Main Street 123",
            "city": "Berlin",
            "zip_code": "10115",
            "country": "Germany"}}},
      "account": {
        "created_at": "2023-08-28T10:15:30Z",
        "status": "active",
        "preferences": {
          "language": "en",
          "currency": "EUR"},
        "subscriptions": {
          "newsletter": true,
          "alerts": {
            "email": true,
            "sms": false}}},
      "meta": {
        "last_login": "2023-08-28T10:20:00Z",
        "login_count": 42,
        "roles": [
          "user",
          "tester"],
        "tags": [
          "load-test",
          "test-user"]}
    }
```
</details>

**Before this improvement**
23 instances for 1000 rps 
max GC pause : 500 µs (high number of small objects generated by the dependency.base functions seems to have contributed here)
max heap objects : 3.17M
max heap usage: 557 MB

**After this improvement**
11 instances for 1000 rps  (50% reduced of CPU power)
max GC pause: 269 µs. 
max heap objects : 1.28M (50% reduced.)
max heap usage: 236 MB 

### Bench-mark test results
_Before_
```BenchmarkAuthorizeRequest/authorize-request-with-body-8                    10002           1068249 ns/op         8432573 B/op             806 allocs/op```

_After_
```BenchmarkAuthorizeRequest/authorize-request-with-body-8                    13461            587883 ns/op         8418624 B/op             566 allocs/op```