@Pushpalanka 2024-08-30
Seeking discussion,

I would like to still **optionally** parse request body only if the filter is configured specifically as `opaAuthorizeRequestWithBody` or  `opaServeResponseWithReqBody` even if a request body is present. Then a user can have a conscious decision to apply the body filters to a narrowed down route where request body is necessary for the policy.

Code currently check for `opa.EnvoyPluginConfig().SkipRequestBodyParse` which seems to be an OPA configuration and I am doubtful if this take into account the filter specification configuration we have as `bodyParsing`

```
func NewOpaAuthorizeRequestWithBodySpec(registry *openpolicyagent.OpenPolicyAgentRegistry, opts ...func(*openpolicyagent.OpenPolicyAgentInstanceConfig) error) filters.Spec {
	return &spec{
		registry:    registry,
		opts:        opts,
		name:        filters.OpaAuthorizeRequestWithBodyName,
		**bodyParsing: true,**
	}
}
```