@MustafaSaber 2024-09-02
maybe the image for the CPU profile after would be great also to visualise the difference