@Pushpalanka 2024-09-02
> maybe the image for the CPU profile after would be great also to visualise the difference

Here we go,

![image](https://github.com/user-attachments/assets/b231273a-21dd-4a26-8b68-12967bb9bfa1)